backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

pydantic==1.10.13
email-validator==2.1.0.post1

firebase-admin==6.4.0

python-dotenv==1.0.1
httpx==0.25.2
python-multipart==0.0.6
